dashboard_automation/pyproject.toml
[project]
name = "dashboard-automation"
version = "0.1.0"
description = "Home Assistant dashboard automation tooling"
authors = [
    {name = "Diego",email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "aiohttp (>=3.13.0,<4.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
